requirements.txt
PyMuPDF
openai
transformers
torch
logging
faiss-cpu
numpy